# Optimize TextEncoder.FindFirstCharacterToEncodeUtf8