@danmosemsft 2020-04-20
@gfoidl is this ready for re-review by @GrabYourPitchforks ?