@gfoidl 2020-04-20
@danmosemsft yes, the requested changes are done.

![grafik](https://user-images.githubusercontent.com/5755834/79727997-8648cf00-82ed-11ea-8e7c-5c2fa3638fb9.png)
Is this "re-request review" something that I could / should have used here?