@gfoidl 2020-06-30
> GitHub auto-linked. :)

Yep, that's really nice as it shows the "rebase diff". Is this OK or shall I still do 
> force-push the original approved commits and then merge origin/master into your branch

?